refactor(auth): refactor authentication flow
